feat(timeline): new check-in card
