+++ src/components/KeyResult/Single/Sections/Timeline/queries.gql
@@ -19,7 +19,7 @@ query GET_KEY_RESULT_TIMELINE_WITH_ID($id: ID!, $first: Int, $after: String) {
             }
             parent {
               confidence
-              progress
+              value
             }
             policy {
               delete
